+++ extension/background.js
@@ -2,6 +2,7 @@ const googleSsoRegex = /name="SAMLResponse" value="([\s\S]+?)"/i;
 const accountSelectionRegex = `tabindex="\\d" jsname="\\S\*" data-authuser="(\\d)" data-identifier="(\\S\*@DOMAIN)"`;
 const stsTokenRegex = /<AccessKeyId>(\S+)<.*\n.*<SecretAccessKey>(\S+)<.*\n.*<SessionToken>(\S+)<.*\n.*<Expiration>(\S+)</i
 const samlFetchErrorRegex = /var problems = {"main": "([\S\s]+)"};/i
+const roleParseRegex = /id="arn:aws:iam::([\S]+)"/
 const googleAccountChooserUrl = 'https://accounts.google.com/AccountChooser'
 const awsSamlUrl = 'https://signin.aws.amazon.com/saml'
 const awsStsUrl = 'https://sts.amazonaws.com'
@@ -22,7 +23,6 @@ const requestHeaders = {
     "Accept-Language": "en-US,en;q=0.9"
 }
 
-
 class portWithExceptions {
     constructor(port) {
         this.postMessage = function (message) {
@@ -30,20 +30,36 @@ class portWithExceptions {
                 port.postMessage(message)
             } catch(err) {
                 console.log(`Error while posting message back to menu. ${err}`)
+            } finally {
+                storage.set({'last_msg_detail':message})
             }
         };
+        this.postError = function (message) {
+            try{
+                port.postMessage(`err: ${message}`)
+                console.error(message)
+            } catch(err) {
+                console.log(`Error while posting message back to menu. ${err}`)
+            } finally {
+                storage.set({'last_msg':'err','last_msg_detail':message})
+            }
+        };        
     }
 }
 
 chrome.runtime.onStartup.addListener(function() {
     storage.get(null, function(props) {
-        refreshAwsTokensInit(props);
+        if (props['autofill']===undefined) storage.set({"autofill":0})
+        if (props['autofill']==1) {
+            awsInit(props, null, 'role_refresh')
+        }
+        if (confCheck(props)) awsInit(props)
     })
 })
 
 chrome.alarms.onAlarm.addListener(function( alarm ) {
     storage.get(null, function(props) {
-        refreshAwsTokensInit(props);
+        awsInit(props);
     })
 });
 
@@ -51,77 +67,149 @@ async function main() {
     chrome.runtime.onConnect.addListener(function(port) {
         let portEx = new portWithExceptions(port);
         port.onMessage.addListener(async function(msg) {
-            console.log('received '+msg)
+            //Stop all background schedule jobs.
             if (msg==='refreshoff'){
-                console.log('turning off background refresh');
-                storage.set({'checked':'0'});
+                storage.set({'checked':0});
                 chrome.alarms.clear("refreshToken");
             }
+            //Start background role refresh
             if (msg==='refreshon')
             {
                 storage.get(null, function(props) {
-                    chrome.alarms.create('refreshToken', { periodInMinutes: parseInt(props.refresh_interval) });
-                    refreshAwsTokensInit(props, portEx);
+                    if(confCheck(props)){
+                        chrome.alarms.create('refreshToken', { periodInMinutes: parseInt(props.refresh_interval) });
+                        awsInit(props, portEx);
+                    } else {
+                        portEx.postError("One or more option isn't configured properly.")
+                    }
                 })
-
+            }
+            //Start role refresh
+            if (msg=='role_refresh') {
+                storage.get(null, function(props){
+                    if (confCheck(props)) awsInit(props, portEx, msg)
+                })
             }
         });
     });    
 }
+
 main()
 
-function refreshAwsTokensInit(props, port=null){
+function confCheck(props){
+    if((props['organization_domain']||props['google_idpid']||props['google_spid']) === ''){
+        return false
+    }
+    return true
+}
+
+function errHandler(port, msg){
+    if (port) {
+        port.postError(msg);
+    } else {
+        console.error(msg);
+        storage.set({'last_msg':'err','last_msg_detail':msg});
+    }
+}
+
+function refreshAwsTokensAndStsCredentials(props,port,samlResponse){
+    let role = props[props.checked]
+    let roleArn=arnPrefix+role
+    let awsAccount=(roleArn.split(":"))[4]
+    let principalArn=`${arnPrefix}${awsAccount}:saml-provider/gsuite`
+    let data = "RelayState=&SAMLResponse="+encodeURIComponent(samlResponse)+"&name=&portal=&roleIndex="+encodeURIComponent(roleArn);
+    fetch(awsSamlUrl, {
+        method: "POST",
+        body: data,
+        headers: requestHeaders
+    }).then(response => response.text())
+    .then((response) => {
+        let errorCheck=response.match(samlFetchErrorRegex)
+        if (errorCheck){
+            let msg = `SAML fetch reponse returned error: ${errorCheck[1]}`
+            errHandler(port, msg)
+        } else {
+            let date = new Date().toLocaleString();
+            console.log(`AWS AlwaysON refreshed tokens successfuly at ${date}`);
+            fetchSts(roleArn, principalArn, samlResponse, props, port)
+        }
+    }).catch((error) => {
+        let msg = `Error in SAML fetch:${error}`
+        errHandler(port, msg)
+    });
+}
+
+function refreshAwsRoles(port,samlResponse){
+    let data = "RelayState=&SAMLResponse="+encodeURIComponent(samlResponse)
+    fetch(awsSamlUrl, {
+        method: "POST",
+        body: data,
+        headers: requestHeaders
+    }).then(response => response.text())
+    .then((response) => {
+        let errorCheck=response.match(samlFetchErrorRegex)
+        if (errorCheck){
+            let msg = `SAML fetch reponse returned error: ${errorCheck[1]}`
+            errHandler(port, msg)
+        } else {
+            let i=0
+            const parseGlobal = RegExp(roleParseRegex, 'g');
+            let matches
+            while ((matches = parseGlobal.exec(response)) !== null) {
+                storage.set({[`role${i}`] : matches[1]})
+                ++i
+            }
+            storage.set({'roleCount': i})
+            if (port) port.postMessage('roles_refreshed')
+
+        }
+    }).catch((error) => {
+        let msg = `Error in SAML fetch:${error}`
+        errHandler(port, msg)
+    });
+}
+
+function awsInit(props, port=null, jobType='refresh'){
     fetch(googleAccountChooserUrl).then(response=> {
         response.text().then(accounts=> {
             var re = new RegExp(accountSelectionRegex.replace("DOMAIN",props.organization_domain),"i");
-            console.log(`Refreshing credentials for ${accounts.match(re)[2]}`)
-            let accountIndex = accounts.match(re)[1]
+            let accountData = accounts.match(re)
+            if(accountData===null){
+                let msg = `Organization domain not found. Please check that you have a Google Account with that domain name logged in.`
+                errHandler(port, msg)
+            }
+            console.log(`Refreshing credentials for ${accountData[2]}`)
+            let accountIndex = accountData[1]
             fetch(`${googleSsoUrl.replace('IDPID',props.google_idpid).replace('SPID',props.google_spid)}${accountIndex}`).then(response => {   
                 response.text().then(result => {
-                    let SAMLReponse=result.match(googleSsoRegex)[1]                    
-                    let role = props[props.checked]
-                    let roleArn=arnPrefix+role
-                    let awsAccount=(roleArn.split(":"))[4]
-                    let principalArn=`${arnPrefix}${awsAccount}:saml-provider/gsuite`
-                    let data = "RelayState=&SAMLResponse="+encodeURIComponent(SAMLReponse)+"&name=&portal=&roleIndex="+encodeURIComponent(roleArn);
-                    fetch(awsSamlUrl, {
-                        method: "POST",
-                        body: data,
-                        headers: requestHeaders
-                    }).then(response => response.text())
-                    .then((response) => {
-                        let errorCheck=response.match(samlFetchErrorRegex)
-                        if (errorCheck){
-                            console.error('Error in saml fetech:', errorCheck[1]);
-                            storage.set({'last_msg':'saml_err'});
-                            if (port) port.postMessage('saml_err');
-                        } else {
-                            let date = new Date().toLocaleString();
-                            console.log(`AWS AlwaysON refreshed tokens successfuly at ${date}`);
-                            fetchSts(roleArn, principalArn,SAMLReponse, props, port)
-                        }
-                    });
+                    let samlResponse=result.match(googleSsoRegex)[1]
+                    switch (jobType) {
+                        case 'role_refresh':
+                            refreshAwsRoles(port, samlResponse)
+                          break;
+                        default:
+                            refreshAwsTokensAndStsCredentials(props, port, samlResponse)
+                    }
                 }).catch((error) => {
-                    storage.set({'last_msg':'sso_process_err'});
-                    if (port) port.postMessage('sso_process_err');
-                    console.error('Error processing sso url:', error);
+                    let msg = `Error processing SSO URL:${error}`
+                    errHandler(port, msg)
                 });
             }).catch((error) => {
-                storage.set({'last_msg':'sso_fetch_err'});
-                console.error('Error fetching sso url:', error);
-                if (port) port.postMessage('sso_fetch_err');
+                let msg = `Error fetching SSO URL:${error}`
+                errHandler(port, msg)
             });
+        }).catch((error) => {
+            let msg = `Error processing Google account chooser data:${error}`
+            errHandler(port, msg)
         });
     }).catch((error) => {
-        storage.set({'last_msg':'account_chooser_err'});
-        console.error('Error finding google account:', error);
-        if (port) port.postMessage('account_chooser_err');
-    });;
+        let msg = `Error finding Google account:${error}`
+        errHandler(port, msg)
+    });
 };
 
-
-function fetchSts(roleArn, principalArn,SAMLReponse, props, port){
-    let STSUrl = `${awsStsUrl}/?Version=2011-06-15&Action=AssumeRoleWithSAML&RoleArn=${roleArn}&PrincipalArn=${principalArn}&SAMLAssertion=${encodeURIComponent(SAMLReponse.trim())}&AUTHPARAMS&DurationSeconds=${props.session_duration}`
+function fetchSts(roleArn, principalArn, samlResponse, props, port){
+    let STSUrl = `${awsStsUrl}/?Version=2011-06-15&Action=AssumeRoleWithSAML&RoleArn=${roleArn}&PrincipalArn=${principalArn}&SAMLAssertion=${encodeURIComponent(samlResponse.trim())}&AUTHPARAMS&DurationSeconds=${props.session_duration}`
     fetch(STSUrl, {
         method: "GET",
         headers: requestHeaders
@@ -139,12 +227,11 @@ function fetchSts(roleArn, principalArn,SAMLReponse, props, port){
               break;
             default:
               stsToken = `export AWS_ACCESS_KEY_ID=${accessKeyId} AWS_SECRET_ACCESS_KEY=${secretAccessKey} AWS_SESSION_TOKEN=${sessionToken} AWS_SESSION_EXPIRATION=${sessionExpiration}`
-          }
+        }
         storage.set({'aws_sts_token':stsToken,'last_msg':'success'});
         if (port) port.postMessage('sts_ready');
     }).catch((error) => {
-        storage.set({'last_msg':'sts_err'});
-        if (port) port.postMessage('sts_err');
-        console.error('Error gettings sts:', error);
+        let msg = `Error getting STS credentials:${error}`
+        errHandler(port, msg)
     });
 }